change in UserRepository interface and PgUserRepository class

+++ src/main/java/Venue_Event_Manager/repository/UserRepository.java
@@ -22,7 +22,7 @@ public interface UserRepository {
 
     List<User> findAllByAccountStatus(Connection conn, AccountStatus accountStatus);
 
-    Optional<Integer> getAverageReview(Connection conn, long userId);
+    Optional<Double> getAverageReview(Connection conn, long userId);
 
     boolean checkPassword(Connection conn, long userId, String password);
